refactor: edit addPost name in controlller folder
