+++ src/controllers/posts/index.ts
@@ -1,6 +1,6 @@
 import getPublicPostsController from './getPublicPostsController';
 import getCohortPostsController from './getCohortPostsController';
-import addPostController from './addPost';
+import addPostController from './addPostController';
 import editPostController from './editPostController';
 import deletePostController from './deletePostController';
 import getSinglePostController from './getSinglePostController';
